refactored apollo config
