+++ components/NavBar/ProfileCorner.tsx
@@ -26,7 +26,7 @@ export default function ProfileCorner() {
   useEffect(() => {
     if (divRef?.current) {
       if (showMenu) {
-        divRef.current.style.display = "block";
+        divRef.current.classList.remove("hidden");
         setTimeout(() => {
           if (divRef.current) divRef.current.style.opacity = "1";
         }, SHOW_MENU_TIME);
@@ -34,7 +34,7 @@ export default function ProfileCorner() {
         divRef.current.style.opacity = "0";
         setTimeout(() => {
           if (!divRef?.current) return;
-          divRef.current.style.display = "none";
+          divRef.current.classList.add("hidden");
         }, SHOW_MENU_TIME);
       }
     }
@@ -73,39 +73,42 @@ export default function ProfileCorner() {
             />
           </button>
           <div
-            className={`none absolute right-1/3 top-16 translate-x-1/2
-            rounded-sm border-[1.8px] border-primary opacity-0 shadow-lg shadow-black/30 
-            duration-200 ease-in-out
-            `}
+            onClick={() => setShowMenu(false)}
+            className="fixed left-0 top-0 z-10 hidden h-full w-full opacity-0 duration-200 ease-in-out"
             ref={divRef}
           >
-            <ul>
-              {profileOptions.map((item) => (
-                <li key={item.path}>
-                  <Link href={`/${item.path}`}>
-                    <div
-                      className="cursor-pointer bg-skin-secondary  px-10
+            <div
+              className={`absolute top-20 right-36 z-20 w-36
+                rounded-sm border-[1.8px] border-primary shadow-lg shadow-black/30`}
+            >
+              <ul>
+                {profileOptions.map((item) => (
+                  <li key={item.path}>
+                    <Link href={`/${item.path}`}>
+                      <div
+                        className="cursor-pointer bg-skin-secondary  px-10
                 py-4 text-sm duration-150 ease-in-out hover:bg-skin-main"
-                    >
-                      <h3 className="text-gray-200">{item.title}</h3>
-                    </div>
-                  </Link>
+                      >
+                        <h3 className="text-gray-200">{item.title}</h3>
+                      </div>
+                    </Link>
+                  </li>
+                ))}
+                <li key={"logout"}>
+                  <button
+                    onClick={async () => {
+                      router.push("/login");
+                      await logOut();
+                      await apolloClient.resetStore();
+                    }}
+                    className="w-full cursor-pointer bg-skin-secondary px-10
+                py-4 text-sm duration-150 ease-in-out hover:bg-skin-main"
+                  >
+                    <h3 className="text-gray-200">logout</h3>
+                  </button>
                 </li>
-              ))}
-              <li key={"logout"}>
-                <button
-                  onClick={async () => {
-                    router.push("/login");
-                    await logOut();
-                    await apolloClient.resetStore();
-                  }}
-                  className="w-full cursor-pointer bg-skin-secondary px-10
-                py-4 text-sm duration-150 ease-in-out hover:bg-skin-main"
-                >
-                  <h3 className="text-gray-200">logout</h3>
-                </button>
-              </li>
-            </ul>
+              </ul>
+            </div>
           </div>
         </div>
       ) : (
