+++ package.json
@@ -26,6 +26,7 @@
     "micro": "^9.4.1",
     "mysql": "^2.18.1",
     "next": "12.2.4",
+    "next-auth": "^4.10.3",
     "next-session": "^4.0.5",
     "pg": "^8.7.3",
     "react": "18.2.0",
